ui/backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-secret-manager
pydantic-settings